Reduced code bloat SIGNIFICANTLY, and also made minor changes to the LLM tone.
