+++ src/app/api/analyze/route.ts
@@ -1,91 +1,21 @@
 import { NextResponse } from "next/server";
-import { jwtVerify } from "jose";
-import crypto from "crypto";
 import OpenAI from "openai";
+import { corsHeaders, handleOPTIONS, verifySession, safeJsonParse } from "../sessionBuilder";
 
 export const runtime = "nodejs";
 
 const openai = new OpenAI({ apiKey: process.env.API_KEY });
 
-const ALLOWED_ORIGINS = new Set([
-  "http://localhost:3000",
-  "https://lifecaddie.org",
-  "https://www.lifecaddie.org"
-]);
-
-function corsHeaders(origin: string | null) {
-  const allowOrigin = origin && ALLOWED_ORIGINS.has(origin) ? origin : "https://lifecaddie.org";
-  return {
-    "Access-Control-Allow-Origin": allowOrigin,
-    "Access-Control-Allow-Methods": "POST,OPTIONS",
-    "Access-Control-Allow-Headers": "Content-Type, Authorization",
-    "Vary": "Origin"
-  };
-}
-
 export async function OPTIONS(req: Request) {
-  return new NextResponse(null, { status: 204, headers: corsHeaders(req.headers.get("origin")) });
-}
-
-function getSecret() {
-  const secret = process.env.LC_SESSION_JWT_SECRET;
-  if (!secret || secret.length < 32) {
-    throw new Error("LC_SESSION_JWT_SECRET must be set and at least 32 characters.");
-  }
-  return new TextEncoder().encode(secret);
-}
-
-function sha256Base64Url(input: string) {
-  return crypto.createHash("sha256").update(input).digest("base64url");
-}
-
-async function verifySession(req: Request) {
-  const auth = req.headers.get("authorization") || "";
-  const token = auth.startsWith("Bearer ") ? auth.slice(7).trim() : "";
-
-  if (!token) return { ok: false as const, reason: "missing_token" };
-
-  try {
-    const { payload } = await jwtVerify(token, getSecret(), {
-      issuer: "lifecaddie",
-      audience: "lifecaddie-space-tool"
-    });
-
-    const ua = req.headers.get("user-agent") || "unknown";
-    const origin = req.headers.get("origin") || "";
-
-    const uaHash = sha256Base64Url(ua);
-    const originHash = origin ? sha256Base64Url(origin) : "";
-
-    if (payload.uaHash !== uaHash) return { ok: false as const, reason: "ua_mismatch" };
-    if (origin && payload.originHash && payload.originHash !== originHash) return { ok: false as const, reason: "origin_mismatch" };
-
-    return { ok: true as const };
-  } catch {
-    return { ok: false as const, reason: "invalid_or_expired" };
-  }
+  return handleOPTIONS(req);
 }
 
 function isAllowedGoal(x: string) {
   return ["moving", "prepping_to_downsize", "reset", "staging", "caregiving", "other"].includes(x);
 }
+
 function isAllowedFeeling(x: string) {
   return ["overwhelmed", "excited", "sad", "motivated", "other"].includes(x);
-}
-
-function safeJsonParse(text: string) {
-  try {
-    return JSON.parse(text);
-  } catch {
-    const a = text.indexOf("{");
-    const b = text.lastIndexOf("}");
-    if (a >= 0 && b > a) {
-      try {
-        return JSON.parse(text.slice(a, b + 1));
-      } catch {}
-    }
-  }
-  return null;
 }
 
 export async function POST(req: Request) {
@@ -127,7 +57,6 @@ export async function POST(req: Request) {
       );
     }
 
-    // Mobile-friendly limits
     if (photo.size > 5_000_000) {
       return NextResponse.json(
         { error: "Photo too large. Please use a smaller image (try under 5MB)." },
@@ -143,7 +72,6 @@ export async function POST(req: Request) {
       );
     }
 
-    // Convert to base64 data URL for the model call
     const arrayBuffer = await photo.arrayBuffer();
     const base64 = Buffer.from(arrayBuffer).toString("base64");
     const dataUrl = `data:${mime};base64,${base64}`;
