+++ src/app/layout.tsx
@@ -1,26 +1,3 @@
-/*
-import './globals.css';
-import { Inter } from 'next/font/google';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata = {
-  title: 'My Next.js App',
-  description: 'Built with the App Router',
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <nav>My Navigation Bar</nav>
-        {children}
-      </body>
-    </html>
-
-    FONTS CAN BE ADDED IN THIS WAY ^
-*/
-
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
