+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-// src/app/space-clarity/page.tsx
 import SpaceClarityTool from "./SpaceClarityTool";
 
 export default function Page() {
